feat: unify composer
