+++ apps/dashboard/src/components/xyne-desk/AppThread/index.ts
@@ -1,3 +1,2 @@
 export { default as AppThread } from './AppThread';
-export { default as AppComposer } from './AppComposer';
 export type { AppDeskMessage } from './AppMessageGroup';
