+++ apps/dashboard/src/components/xyne-desk/DeskComposer/index.ts
@@ -1,0 +1,2 @@
+export { default as DeskComposer } from './DeskComposer';
+export type { DeskComposerVariant } from './DeskComposer';
