--- apps/dashboard/src/components/xyne-desk/DeskReplyComposer/index.ts
@@ -1,1 +1,0 @@
-export { SocialMediaReplyComposer } from './SocialMediaReplyComposer';
