+++ apps/dashboard/src/components/xyne-desk/SlackThread/index.ts
@@ -1,4 +1,3 @@
 export { default as SlackThread } from './SlackThread';
-export { default as SlackComposer } from './SlackComposer';
 export { default as SlackMessage } from './SlackMessage';
 export type { SlackEmailMessage } from './SlackMessage';
